requirements.txt
streamlit
boto3
pytesseract
pdf2image
Pillow